Add price display to InfoComponent:
Added display of the paid version price in the UI by passing the formatted Price value from the TestDTO to the InfoComponent.
Changes made in ArcadeGame.razor to pass the paid version price to the InfoComponent.
Enhance InfoComponent layout:

Updated the layout of the InfoComponent to include detailed information about the paid version, with icons representing the different features of the premium version (duration, number of questions, progress tracking, and additional features).
The price of the premium version was also added to the "Get Premium" button.
New property for Price in TestDTO:

A new Price property was added to TestDTO to store the price of the paid version (in cents). The value is retrieved and used in the components.